add resized cards, fully automatic vec, dat generator and trainer scripts, removing unused bat-s

+++ resources/learning_resources/train.bat
@@ -1,11 +1,8 @@
 @echo off
-set posNumber=%1
-set haarDir=%2
-set width=%3
-set height=%4
-set minhitrate=%5
-set maxfalsealarm=%6
-if exist %haarDir%    (
-	rd  /S/Q %haarDir%  
+set vecDirectory=./VECs/
+set haarDir=./XMLs/
+
+for /F "tokens=*" %%A in (imagesToProcess.txt) do (
+if exist %haarDir%%%A  rd  /S/Q %haarDir%%%A  
+opencv_haartraining -data %haarDir%%%A  -vec %vecDirectory%%%A_samples.vec  -bg  negatives.dat -nstages 12 -nsplits 2 -minhitrate 0.999 -maxfalsealarm 0.2 -npos 250 -nneg 125 -w 40 -h 58 -nosym -mem 2048 -mode ALL
 )
-opencv_haartraining -data %haarDir%   \ -vec samples4.vec -bg  negatives.dat -nstages 12 -nsplits 2 -minhitrate %minhitrate% -maxfalsealarm %maxfalsealarm% -npos %posNumber% -nneg 125 -w %width% -h %height% -nosym -mem 2048 -mode ALL
